change Bifurcator system ID

+++ src/engine/platform/bifurcator.cpp
@@ -54,7 +54,7 @@ void DivPlatformBifurcator::acquire(short** buf, size_t len) {
     for (int i=0; i<4; i++) {
       chan[i].audSub+=chan[i].freq;
       if (chan[i].audSub>=65536) {
-        long long newx=(long long)chan[i].curx*(chan[i].param+65536)/32768;
+        int64_t newx=(int64_t)chan[i].curx*(chan[i].param+65536)/32768;
         newx*=65536-chan[i].curx;
         chan[i].curx=(int)(newx/65536);
         chan[i].audSub&=65535;
